Extract notifications and messages to a service
